Moving from inline scripts to script files and ansible
Ansible is now used for configuration of the VM using the local connection mode.

A set of playbooks are uploaded automatically by packer and then executed.

Shell scripts are needed to install ansible and needed packages and again to remove ansible and cleanup the system.

This will mean supporting new versions, OSes etc. will be much easier and better organised.