Adjusting shutdown command to make a hidden script
More of a cosmetic change but the shutdown command packer uses creates a shutdown script in the vagrant home directory, then executes it (shutting down the VM).

Since this is the last thing to be executed before the machine shutdown I don't think there's a way to easily delete this file, so its left behind.

To make the home directory look clean this file is now hidden .