Finally fixing the update/upgrade problem
The provisioner now reboots the VM after running apt-get upgrade meaning upgraded packages are in use *before* we install virtualbox guest additions and therefore won't break when used as a vagrant box later on.

This is sustainable as creating a new release simply requires re-running the packer task (which is one line) and won't need to be done that often.