Remove non-supported base boxes
This drastically reduces the scope of this project to supporting just one base box (ubuntu).

We can't commit to supporting anymore than this and their inclusion in this project was misleading as we never actually released any others successfully. If other users want different boxes they can either contribute to this repository and release boxes as needed, or use boxes provided by someone else (preferable).