Renaming to reflect new stable purpose
We can confidently support the single base box we now support and the practices for creating and deploying this box are now stable and proved to work.

Though improvements to the building process (though merging changes from the 'upstream' bento project) still need to be made, that will be a recurring process so has no bearing on moving this project from an experiment to 'stable'.