OVA files need to made manually for VMware
For some reason (hopefully a bug) the ovftool will not produce a complete manifest for OVA files, but will for OVF packages.

The manifest is simply a list of SHA1 hashes for files within a OVF/OVA package. It isn't needed but VMware will complain if its missing or incomplete (including ESXI).

Producing an OVF adds manifest entries for both the virtual hard disk and OVF meta-data file. Producing an OVA only includes an entry for the virtual hard disk, despite the OVF file still being there.

Converting from an OVF to OVA removes this second manifest entry, i.e. it actively removes it since the OVF has both.

To workaround this its possible to convert an OVF package to an OVA file manually. The steps to do this are now listed in the README.

This adds some complexity but if you follow the steps there shouldn't be any problems. We now explictly check the OVA files we produce validate against the format, which is an improvement regardless.