Adding support for SSH agent forwarding whilst using sudo
This support was previously configured during provisioning, however since it is standard it is better we add this support here.

For other environments where this base image is not used (i.e. cloud providers) templates should be created containing this change, or support should be added within their bootstrap provisioning roles as required.