Adding proof of concept CentOS template
This still requires tweaking to allow the Agent Forwarding environment variable through to Sudo.

It will however produce a minified build of CentOS 7.1 for both VirtualBox and VMware with working shared folders support. Changes were required to sevral of the setup scripts as these seemed to be erroring when there was no error!

This commit adds an 'alpha' of 0.1.0 for this template. Its expected there will be a 'beta' and then 0.1.0 which will be listed publicly (after being tested by PDC).