Enabling system firewall on CentOS by default
This is a conventional default for CentOS anyway.

As this is technically a breaking change, but before the 1.0.0 release a new minor release is needed (0.3.0).

On desktops, KickStart can be configured to enable the firewall and allow SSH.
On cloud images, a script is used to achieve the same result.