Adding template information to templates
This is to facilitate looking up the template used in an OS, specifically its name and version.

Currently these are expressed in a form Ansible can read as a local fact, however the actual file they are written to is simply an INI file so should be usable for other purposes as well (just from a non-obvious location).

This is a new feature meaning new minor releases have been made for both templates.