Adding host key regeneration support
Previously, all non-cloud instances of templates used the same host key - which is obviously bad.

Fixing this is relatively simple, on Ubuntu a command is added to regenerate missing keys on boot (i.e. first boot) - on CentOS this happens automatically - then, as a final provisioning step, remove the existing host keys.

As this is backwards compatible, a new minor releases is needed.