Forcing install of VMware tools
This is a workaround as VMware have changed the installation defaults to not install VMware tools by default - which of course broke shared folders.

They are now recommending we use openVM tools or similar - but their knowledge base is currently down and support for shared folders is not included so for know things will stay as they are.