Adding conventional 'terraform' user to cloud artefacts
As the 'vagrant' user is to OVA based artefacts, so the 'terraform' user will be to cloud based artefacts. This is mainly useful for automated provisioning to mask differences between providers and operating systems.

A public key for a per-provider identity is used as an authorized_key for this new user.

These changes are an addition and released as a minor version.